refactor: test message creators
