+++ tests/utils/__init__.py
@@ -1,2 +1,3 @@
-from .creators import *
+from .messages.creators import *
+from .messages.creators.constants import *
 from .mocks import *
